requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.4
filelock==3.16.1
imageio==2.36.1
joblib==1.4.2
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tifffile==2024.12.12
tzdata==2024.2
virtualenv==20.28.1
